move preview elements to their own component
