+++ app/ui/components/Header.tsx
@@ -1,8 +1,8 @@
 "use client"
-import { Button } from "@/app/ui/inputs/Button"
 import Nav from "@/app/ui/components/Nav"
 import { Logo } from "@/app/ui/components/Logo"
 import dynamic from "next/dynamic"
+import Link from "next/link"
 
 const SlIcon = dynamic(
   () => import("@shoelace-style/shoelace/dist/react").then(mod => mod.SlIcon),
@@ -15,10 +15,13 @@ export function Header() {
     <header className="flex justify-between items-center self-stretch bg-white pl-6 pr-4 py-4 rounded-xl mb-4 md:mb-6">
       <Logo />
       <Nav></Nav>
-      <Button variant="secondary">
+      <Link
+        href="/preview"
+        className="rounded-lg font-semibold text-base border-purple text-purple active:bg-light-purple disabled:opacity-25 bg-white border border-solid px-4 py-[11px] md:px-[27px]"
+      >
         <SlIcon name="eye" className="h-5 w-5 sm:hidden" />
         <span className="hidden sm:inline">Preview</span>
-      </Button>
+      </Link>
     </header>
   )
 }
